requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
pydantic==2.10.6
requests==2.32.3
flask
dotenv
pydantic[email]
python-dateutil